+++ camera/src/commonMain/kotlin/tech/kotlinlang/camera/CameraHelper.kt
@@ -1,0 +1,9 @@
+package tech.kotlinlang.camera
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+interface CameraHelper {
+    @Composable
+    fun CameraContent(modifier: Modifier)
+}