+++ camera/src/commonMain/kotlin/tech/kotlinlang/camera/Util.kt
@@ -1,0 +1,4 @@
+package tech.kotlinlang.camera
+
+expect fun getCameraHelper(): CameraHelper
+