+++ camera/src/iosMain/kotlin/tech/kotlinlang/camera/IosCameraHelper.kt
@@ -1,0 +1,10 @@
+package tech.kotlinlang.camera
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+class IosCameraHelper: CameraHelper {
+    @Composable
+    override fun CameraContent(modifier: Modifier) {
+    }
+}