+++ composeApp/src/commonMain/kotlin/tech/kotlinlang/permission/App.kt
@@ -2,8 +2,11 @@ package tech.kotlinlang.permission
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBars
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBars
@@ -21,10 +24,12 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.ui.tooling.preview.Preview
+import tech.kotlinlang.camera.getCameraHelper
 import tech.kotlinlang.permission.location.LocationRequestResult
 import tech.kotlinlang.permission.result.CameraPermissionResult
 import tech.kotlinlang.permission.result.LocationPermissionResult
@@ -107,6 +112,14 @@ private fun CameraPermissionContent(
 
             CameraPermissionResult.Granted -> {
                 Text("Camera Permission is Granted")
+                val cameraHelper = remember { getCameraHelper() }
+                Spacer(Modifier.height(12.dp))
+                cameraHelper.CameraContent(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .aspectRatio(1F)
+                        .clip(RoundedCornerShape(20.dp))
+                )
             }
         }
     }